pmsrv-service: ensure pmsrv is loaded
New meta-owasys layer upgrade will change
the default order the power-manager service is loaded.
To have it available for BalenaOS we need to ensure
it is started at the very beginning of the system boot.
Otherwise BalenaOS will have problems when trying to bring up
both wifi and cellular interfaces as they are powered up
by this pmsrv.service

Changelog-entry: power-manager is loaded properly
Signed-off-by: Alvaro Guzman <alvaro.guzman@owasys.com>